scripts/requirements_v255.txt
# Dive AI V25.5 Requirements
# All 20 voice techniques + V98 API + fixes

# Core Dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
scipy>=1.11.0

# Voice Processing (Techniques #1-4, #7, #11, #13, #18)
torch>=2.0.0
torchaudio>=2.0.0
silero-vad>=5.0.0
openai-whisper>=20231117
pyttsx3>=2.90
pygame>=2.5.0
pyaudio>=0.2.13

# Wake Word Detection (#7)
openwakeword>=0.5.0
pvporcupine>=3.0.0

# Audio Processing (#8, #9, #12, #19)
dtln-aec>=1.0.0
rnnoise-python>=1.0.0
webrtc-noise-gain>=1.0.0
sounddevice>=0.4.6

# Semantic Caching (#5)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Vision (#14)
pillow>=10.0.0
opencv-python>=4.8.0
mss>=9.0.1

# Monitoring & Resilience (#16, #17)
psutil>=5.9.0
tenacity>=8.2.0

# Web/Desktop (#6)
flask>=3.0.0
flask-cors>=4.0.0
websockets>=12.0

# Utilities
colorama>=0.4.6
tqdm>=4.66.0
pydantic>=2.0.0
